EPA Air Quality Monitoring Sites Metadata

This dataset contains metadata for air quality monitoring sites in the United States. It includes site location details, land use classification, elevation, operational dates, and other relevant attributes.

Usage

epa_airdata_monitoring_sites

Format

A tibble with the following columns:

State Code

FIPS code representing the U.S. state.

County Code

FIPS code representing the county within the state.

Site Number

Unique identifier for the monitoring site within the county.

Latitude

Latitude of the monitoring site in decimal degrees.

Longitude

Longitude of the monitoring site in decimal degrees.

Datum

Geodetic datum used for latitude/longitude coordinates (e.g., "WGS84").

Elevation

Elevation of the site in meters above sea level.

Land Use

Classification of land use surrounding the site (e.g., "RESIDENTIAL", "AGRICULTURAL").

Location Setting

General setting of the site (e.g., "URBAN AND CENTER CITY", "RURAL").

Site Established Date

Date when the monitoring site was established.

Site Closed Date

Date when the site ceased operation (if applicable).

Met Site State Code

FIPS code of the meteorological site, if applicable.

Met Site County Code

FIPS code of the county where the meteorological site is located, if applicable.

Met Site Site Number

Unique identifier of the associated meteorological site, if applicable.

Met Site Type

Type of meteorological site, if applicable.

Met Site Distance

Distance from the monitoring site to the associated meteorological site, in kilometers.

Met Site Direction

Cardinal direction from the monitoring site to the meteorological site.

GMT Offset

Offset from GMT (Greenwich Mean Time) in hours.

Owning Agency

Agency responsible for maintaining the monitoring site.

Local Site Name

Locally assigned name of the monitoring site.

Address

Physical address of the monitoring site.

Zip Code

ZIP code where the monitoring site is located.

State Name

Full name of the state where the site is located.

County Name

Full name of the county where the site is located.

City Name

City where the monitoring site is located, if applicable.

CBSA Name

Core-Based Statistical Area (CBSA) associated with the site, if applicable.

Tribe Name

Name of the Native American tribe associated with the site, if applicable.

Extraction Date

Date when this data was extracted.

Source

https://www.epa.gov/outdoor-air-quality-data

Examples

data(epa_airdata_monitoring_sites)
head(epa_airdata_monitoring_sites)
#> # A tibble: 6 × 28
#>   `State Code` `County Code` `Site Number` Latitude Longitude Datum Elevation
#>   <chr>        <chr>         <chr>            <dbl>     <dbl> <chr>     <dbl>
#> 1 01           001           0001              32.4     -86.5 WGS84        64
#> 2 01           001           0002              32.4     -86.4 WGS84         0
#> 3 01           001           0003              32.3     -86.8 WGS84        41
#> 4 01           003           0001               0         0   NAD27         0
#> 5 01           003           0002              30.6     -87.7 WGS84         0
#> 6 01           003           0003              30.6     -87.7 WGS84        49
#> # ℹ 21 more variables: `Land Use` <chr>, `Location Setting` <chr>,
#> #   `Site Established Date` <date>, `Site Closed Date` <date>,
#> #   `Met Site State Code` <chr>, `Met Site County Code` <chr>,
#> #   `Met Site Site Number` <chr>, `Met Site Type` <chr>,
#> #   `Met Site Distance` <dbl>, `Met Site Direction` <chr>, `GMT Offset` <dbl>,
#> #   `Owning Agency` <chr>, `Local Site Name` <chr>, Address <chr>,
#> #   `Zip Code` <chr>, `State Name` <chr>, `County Name` <chr>, …